Successfully ran spi with dma on registers. Notices some addresses shift. Current speed is 120kS/s

+++ Radar_controller_example/Example_Any_Frequency/Core/Src/utils.c
@@ -44,7 +44,7 @@ void* LightLED(void* state){
 
 void* ReadADC(void* samples){
 	static bool ret;
-	uint16_t* value = (uint16_t*)samples;
+	uint32_t* value = (uint16_t*)samples;
 	if (*value <= 0 && *value > ad7676_data->data_ptr_max) ret = false;
 	else {
 		ad7676_read_samples(*value);
